Refactor subject browse container and components
- Migrate components from simplified single page component to individual components and elements
- Basic styling set up
