+++ frontend/src/App.js
@@ -1,7 +1,8 @@
 import "./assets/fonts.css";
 import "./App.css";
 import ArticleList from "./components/ArticleList/ArticleList";
-import Subjects from "./containers/Subjects/Subjects";
+import SubjectBrowse from "./containers/Subjects/SubjectBrowse";
+import Simple from "./containers/Simple/Simple";
 import Article from "./containers/Article/Article";
 import Homepage from "./containers/Homepage/Homepage";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
@@ -16,7 +17,7 @@ function App() {
               <Homepage />
             </Route>
             <Route exact path="/subjects">
-              <Subjects />
+              <SubjectBrowse />
             </Route>
             <Route path="/articles">
               <ArticleList />
